Work done so far
MainPage.html:
The div element with nothing in between is most likely going to be our side bar for viewing our playists

On the createPlaylistForm, there is a little of JS tied in there. It just does the pop up type of animation when the create a playlist button gets clicked

mainPageSheet:
mostly has the div box element for our form.

mainPageJS:
The code there is just to open the form and close the form.